pyproject.toml
[project]
name = "dados-bcb"
version = "0.1.0"
description = "Coleta e gestao de dados economicos do BCB e outras fontes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastparquet>=2024.11.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "py7zr>=0.20.0",
    "pyarrow>=22.0.0",
    "python-bcb>=0.3.3",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bacen", "src/data", "src/mte"]
